requirements.txt
alembic==1.4.2
atomicwrites==1.4.0
attrs==19.3.0
cffi==1.14.0
click==7.1.2
colorama==0.4.3
coverage==5.2
cryptography==3.0
Flask==1.1.2
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
more-itertools==8.4.0
mysqlclient==1.3.12
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.0
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.18
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.1.0
